Backend/WhatMunch_Django/requirements.txt
asgiref
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
PyJWT
pytz
sqlparse
psycopg2-binary
python-dotenv
django-allauth
django-allauth[socialaccount]
python-decouple
pytest
pytest-django
uvicorn
whitenoise
